refactor(cmd/mdsmith): extract backlink algorithm into internal/backlinks
The 2026-08-30 architecture audit found ~430 of backlinks.go's 585
lines carrying the backlink target-matching algorithm (link/wikilink
resolution, workspace-relative path math) inside the cmd/mdsmith CLI
package, violating go.md's "domain logic belongs in pkg/mdsmith,
internal/engine, or their dependencies" rule.

Extract Record, Collect, and every private helper the matching
algorithm needs into a new internal/backlinks package.
cmd/mdsmith/backlinks.go now only parses flags, validates arguments,
calls backlinks.Collect, and formats output. No behavior change:
existing unit tests moved to internal/backlinks/backlinks_test.go
(CLI-layer tests for flag parsing, argument validation, and output
formatting stay in cmd/mdsmith), and the e2e test is untouched.

Also records the full 2026-08-30 audit sweep in the architecture
audit log, files plans for the two remaining tax findings (missing
unit tests for WordFrequencyInto/directivefiles helpers; RunCache's
package placement), and shuffles the audit-log archive shards to
stay under the file-length budget.
