+++ docs/development/architecture-audit-archive.md
@@ -21,85 +21,11 @@ on to
 It hit the budget again on 2026-08-23; entries
 from 2026-05-13 through 2026-05-19 moved on to
 [the second archive](architecture-audit-archive-2.md)
-as well.
-
-## Audit 2026-05-31 (range: 4809097..37488a7)
-
-Plans 200, 201, 202 green. Tax:
-[plan/223][223] (`pkg/mdsmith` private helpers),
-[plan/224][224] (`internal/lint` SRP, now 12 files).
-`linkstyle` helpers — add tests in-place.
-
-[223]: ../../plan/223_arch-fix-mdsmith-helper-tests.md
-[224]: ../../plan/224_arch-fix-lint-srp.md
-
-### nice-to-have (2026-06-02)
-
-`internal/punkt` not in the layering map —
-[plan/225][225]. Separately, [plan/224][224]
-(`internal/lint` SRP) is now implemented:
-`gitignore`, `bytelimit`, and `piparser`
-split into sibling packages.
-
-[225]: ../../plan/225_arch-fix-punkt-layering.md
-
-## Audit 2026-06-07 (range: 37488a7..82583fc)
-
-Plans 203–225 green. Blocker: `Session.CheckSource`
-(public API) had no unit test. Fixed: added
-`pkg/mdsmith/checksource_test.go` with 4 tests.
-Tax: the [tablereadability dedup][2606071930] and
-[include helper test][2606071931] plans.
-
-[2606071930]: ../../plan/2606071930_arch-fix-tablereadability-dup.md
-[2606071931]: ../../plan/2606071931_arch-fix-include-helper-tests.md
-
-## Audit 2026-06-14 (range: 82583fc..aed18aa)
-
-Tax: [build→rules DIP](../../plan/2606141910_arch-fix-build-rules-dip.md),
-[engine wrappers](../../plan/2606141911_arch-fix-engine-deprecated-wrappers.md),
-[secreview tests](../../plan/2606141912_arch-fix-secreview-report-tests.md).
-
-## Audit 2026-06-16 (range: aed18aa..7793b97)
-
-Lazy-parse series (plans 2606141901–2606141904).
-Tax: [new-pkg-docs](../../plan/2606162213_arch-fix-new-pkg-docs.md),
-[helper-tests](../../plan/2606162214_arch-fix-missing-helper-tests.md).
-
-## Audit 2026-06-21 (range: 7793b97..e701b94)
-
-Parity + Layer-0 parse-skip series.
-Symlink containment; engine panic recovery.
-VS Code `kinds` and `rule-doc` commands.
-270 Go/TS sources. No blockers,
-rule-to-rule imports, or DIP violations.
-
-### tax (2026-06-21)
-
-- `internal/engine/runner.go` (1 290 lines) —
-  SRP: 7 concerns. Fixed this cycle: split into
-  `runner_layer0.go`, `runner_cache.go`,
-  `runner_log.go` — [plan/2606211907][2606211907].
-
-- `internal/lint/layer0.go` (1 203 lines) —
-  full Layer-0 block scanner. Fix: split along
-  block-type sub-parsers —
-  [plan/2606211908][2606211908].
-
-- `internal/lsp/server.go` (1 007 lines) —
-  crept back over 1 000 lines. Dispatch-group
-  split — [plan/2606211909][2606211909].
-
-### nice-to-have (2026-06-21)
-
-- `pkg/mdsmith/workspace.go` trivial methods
-  lack "// no test by design" exemptions —
-  [plan/2606211910][2606211910].
-
-[2606211907]: ../../plan/2606211907_arch-fix-runner-srp-split.md
-[2606211908]: ../../plan/2606211908_arch-fix-layer0-split.md
-[2606211909]: ../../plan/2606211909_arch-fix-lsp-server-split.md
-[2606211910]: ../../plan/2606211910_arch-fix-workspace-exemptions.md
+as well. It hit the budget again on 2026-08-30;
+entries from 2026-05-31 through 2026-06-21 moved
+on to
+[the second archive](architecture-audit-archive-2.md)
+too.
 
 ## Audit 2026-06-23 (range: e701b94..1599c9f)
 
@@ -271,3 +197,88 @@ None.
 
 [tests]: architecture/tests.md
 [go]: architecture/go.md
+
+## Audit 2026-08-09 (range: 6680ff5..2ab4b29)
+
+100 commits, 148 files touched (92 Go files).
+
+New packages this cycle:
+
+- `internal/foreignregion` — foreign managed-region
+  protection for check/fix.
+- `internal/convention` — extracted convention model.
+
+New opt-in rules this cycle:
+
+- `internal/rules/occurrence` (MDS060).
+- `internal/rules/slidevstructure` (MDS073).
+
+No rule-to-rule imports. No reverse-layer imports. No
+Liskov breaks. `cmd/mdsmith/main.go` (709 lines) stayed well
+under the ~1000-line threshold, as did
+`internal/lsp/server.go`/`symbols.go` (554/511 lines).
+
+Clean surfaces, verified:
+
+- `internal/mdpath` consolidation: `IsMarkdownPath` grew into
+  a single source of truth (`Extensions`, `FileGlobs`,
+  `RecursiveGlobs`, `HasMarkdownExt`) now consumed by
+  `internal/lsp`, `internal/config`, and the merge-driver
+  include set — SRP fix per go.md's "package answers one
+  question." Every function has an exact-name test.
+- `cmd/mdsmith/init.go` split: extracted from `main.go` with
+  full dedicated-test coverage — 39 `Test*` functions cover
+  all 13 functions, including error paths.
+- `internal/rules/crossfilereferenceintegrity`'s
+  double-checked-locking memoization
+  (`cachedGlobSettingsErr`) is pinned by dedicated tests
+  following the `TestReceiver_Foo` naming convention.
+
+### blockers (2026-08-09)
+
+- `MDS073` is claimed by two unrelated, independently
+  registered diagnostics. `internal/rules/slidevstructure/rule.go:52`
+  returns `"MDS073"` for the Slidev slide-structure rule
+  (registered in `internal/rules/all/all.go`, documented at
+  `internal/rules/MDS073-slide-structure/README.md`).
+  `internal/foreignregion/scan.go:25` independently defines
+  `RuleID = "MDS073"` for the malformed-marker-pair
+  diagnostic (wired into `internal/engine/runner.go` and
+  `internal/fix/fix.go`, documented at
+  `docs/reference/foreign-regions.md:76`). Both landed the
+  same day (2026-07-19) — slide-structure was renumbered
+  `MDS072 -> MDS073` in `7e4925a` while the foreign-region
+  feature independently claimed `MDS073` in parallel commits
+  — an undetected rebase/merge collision.
+  [cross-system.md][cross] treats rule IDs as part of the
+  public `.mdsmith.yml`/CLI/docs contract; no test in
+  `internal/integration/` or `internal/rule/` enforces
+  uniqueness across `rule.All()` plus out-of-band IDs like
+  foreign-region's. Fixed by renumbering the foreign-region
+  diagnostic to the next free ID and adding a uniqueness
+  contract test — [plan/2608091910][2608091910].
+
+### tax (2026-08-09)
+
+- `internal/rules/occurrence/rule.go` has 18 non-trivial
+  private helpers with no dedicated unit test by name
+  (`countCombinedInRange`, `searchText`, `applyScope`, …).
+  All are covered indirectly by 47 behavior-level
+  `TestCheck_*`/`TestApplySettings_*` cases, but
+  [tests.md][tests] requires a test by the function's own
+  name, and the pre-existing sibling rule
+  `paragraphstructure` follows that convention for every
+  helper — [plan/2608091911][2608091911].
+- `internal/rules/slidevstructure/rule.go` has 11
+  non-trivial private helpers (`checkSlide`, `slideAnchor`,
+  `nearest`, …) in the same shape — covered behaviorally by
+  35 `Test*` cases but not by name —
+  [plan/2608091911][2608091911].
+
+[cross]: architecture/cross-system.md
+[2608091910]: ../../plan/2608091910_arch-fix-mds073-collision.md
+[2608091911]: ../../plan/2608091911_arch-fix-missing-unit-tests.md
+
+### nice-to-have (2026-08-09)
+
+None found this cycle beyond the tax items above.
